refactor: modularize quotes and favorites
